refactor(vs-component): moved drra.rs to a module called 'models'; updated links to it
