+++ modules/vs-component/src/models.rs
@@ -1,0 +1,1 @@
+pub mod drra;
