+++ modules/vs-component/src/utils.rs
@@ -1,4 +1,4 @@
-use crate::drra::ParameterList;
+use crate::models::drra::ParameterList;
 use bs58::encode;
 use log::{debug, warn};
 use serde::Serialize;
